Merge-in from VC8 sandbox - Pixelbuffers for DX7 (DynTex runs!) - getBuffer(face,mip) and getPixelBox(face,mip) now take size_t instead of int, as negative values make no sense - Add OgreHardwarePixelBuffer::blit convience function to blit entire contents of source PixelBuffer - PixelUtil::getBitMasks to query masks on native endian formats

+++ OgreMain/include/OgreTexture.h
@@ -223,7 +223,7 @@ namespace Ogre {
 			@remarks	The buffer is invalidated when the resource is unloaded or destroyed.
 						Do not use it after the lifetime of the containing texture.
 		*/
-		virtual HardwarePixelBufferSharedPtr getBuffer(int face=0, int mipmap=0) = 0;
+		virtual HardwarePixelBufferSharedPtr getBuffer(size_t face=0, size_t mipmap=0) = 0;
 
     protected:
         unsigned long mHeight;
